requirements.txt
requests
python-dotenv
psycopg2-binary
streamlit
pandas
sqlalchemy
watchdog
